requirements.txt
streamlit-audiorec
gtts
openai
clarifai_grpc
Pillow
numpy
matplotlib